Update link in expo-notifications.json
Doc at "https://docs.expo.dev/versions/latest/sdk/notifications/#api" has a broken link for `TaskManager.defineTask` to "/task-manager#taskmanagerdefinetasktaskname-taskexecutor". I'm not entirely sure how redirects are supposed to work on the site but that gives me a 404. The rest of the doc seems to use links that link to "/versions/latest/sdk/...", so hopefully that's the right way to do it... changed it to "versions/latest/sdk/task-manager/#taskmanagerdefinetasktaskname-taskexecutor"